Moved Deconvolution Test to a development project
