+++ mzLib/Test/TestDeconvolution.cs
@@ -12,7 +12,7 @@ using System.Globalization;
 using System.IO;
 using System.Linq;
 
-namespace Test.DeconvolutionTests
+namespace Test
 {
     [TestFixture]
     [ExcludeFromCodeCoverage]
